remove float from rendering process
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@38693 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/geogebra/common/geogebra3D/euclidian3D/openGL/GLBuffer.java
@@ -26,7 +26,7 @@ public interface GLBuffer {
 	 * put value at current buffer position
 	 * @param value value
 	 */
-	public void put(float value);
+	public void put(double value);
 	
 
 	/**
@@ -37,7 +37,7 @@ public interface GLBuffer {
 	 * @param length
 	 *            length to copy
 	 */
-	public void set(ArrayList<Float> array, int length);
+	public void set(ArrayList<Double> array, int length);
 
 	/**
 	 *
